fix(circular-import): Fix circular import errors largely surrounding the alias module. Extract alias module out of utils subpackage into ast package where it is used.
