+++ fhy/lang/ast/passes/identifier_collector.py
@@ -11,8 +11,8 @@ Classes:
 from typing import Set
 
 from fhy import ir
+from fhy.lang.ast.alias import ASTObject
 from fhy.lang.ast.visitor import Visitor
-from fhy.utils.alias import ASTObject
 
 
 class IdentifierCollector(Visitor):
